Split service need into pieces
Experimental change aiming for less clutter
