+++ frontend/packages/citizen-frontend/src/localization/fi.ts
@@ -227,7 +227,8 @@ export default {
             agreed:
               'Olemme yhdessä sopineet lapsen esiopetuksen hakemisesta lomakkeen tietojen mukaisesti.',
             notAgreed:
-              'Emme ole voineet sopia hakemuksen tekemisestä yhdessä / toisella huoltajalla on vain tiedonsaantioikeus.'
+              'Emme ole voineet sopia hakemuksen tekemisestä yhdessä / toisella huoltajalla on vain tiedonsaantioikeus.',
+            noAgreementStatus: 'Ei tiedossa'
           },
           fridgePartner: {
             title:
