+++ frontend/packages/citizen-frontend/src/localization/sv.ts
@@ -232,7 +232,8 @@ const sv: Translations = {
             agreed:
               'Olemme yhdessä sopineet lapsen esiopetuksen hakemisesta lomakkeen tietojen mukaisesti.',
             notAgreed:
-              'Emme ole voineet sopia hakemuksen tekemisestä yhdessä / toisella huoltajalla on vain tiedonsaantioikeus.'
+              'Emme ole voineet sopia hakemuksen tekemisestä yhdessä / toisella huoltajalla on vain tiedonsaantioikeus.',
+            noAgreementStatus: 'Ei tiedossa'
           },
           fridgePartner: {
             title:
